+++ src/fit.jl
@@ -60,6 +60,8 @@ function fit_mle(
     error_group_cap::Integer=8,
     error_eta::Union{Nothing,Real,Symbol}=nothing,
     edge_index::Union{Nothing,AbstractVector{<:Integer}}=nothing,
+    error_blocks::Union{Nothing,Symbol}=nothing,
+    firm_group::Union{Nothing,AbstractVector{<:Integer}}=nothing,
     solver::AbstractGMRFSolver=ExactCholesky(),
     rho_limit::Union{Real,Symbol}=0.99,
     fix_rho::Union{Nothing,Float64}=nothing,
@@ -80,6 +82,8 @@ function fit_mle(
         error_group_cap=error_group_cap,
         error_eta=error_eta,
         edge_index=edge_index,
+        error_blocks=error_blocks,
+        firm_group=firm_group,
     )
     return fit_mle(M, ss;
         solver=solver,